backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["nnnLik <rslmakhmudov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.115.3"
uvicorn = "0.32.0"
psycopg = {extras = ["binary"], version = "3.2.3"}
pydantic = "2.9.2"
pydantic-settings = "2.6.0"
email-validator = "2.2.0"
ipython = "8.28.0"
passlib = "1.7.4"
pyjwt = "2.9.0"
python-multipart = "0.0.16"
aiokafka = {extras = ["lz4"], version = "0.12.0"}
orjson = "3.10.10"
punq = "0.7.0"
sqlalchemy = "2.0.36"
asyncpg = "^0.30.0"

[tool.poetry.group.dev.dependencies]
ruff = "0.7.0"
pytest = "8.3.3"
mypy = "1.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
